requirements.txt
streamlit
requests
langchain-google-genai
langchain-core
python-dotenv
wikipedia-api
watchdog
setuptools